:recycle: Refactor: 임베딩 모델을 KoBERT에서 KURE-v1으로 교체
- monologg/kobert에서 nlpai-lab/KURE-v1로 임베딩 모델 변경
- kobert.py에서 embedding_model.py로 파일명 및 클래스명(EmbeddingModel) 변경
- 모델 특성 따라 CLS 토큰 풀링에서 mean pooling 방식으로 변경
- 임베딩 벡터 차원 768에서 1024로 변경 (models.py, 마이그레이션 파일)
