+++ ai-dispatcher/app/service/embedding_service.py
@@ -1,19 +1,19 @@
 import numpy as np
 import torch
-from app.model.kobert import KoBERTModel
+from app.model.embedding_model import EmbeddingModel
 
 def extract_embedding(text: str) -> np.ndarray:
     """
-    공고 제목 텍스트를 받아서 KoBERT로 768차원 임베딩 벡터를 추출해 반환
+    공고 제목 텍스트를 받아서 KURE-v1으로 1024차원 임베딩 벡터를 추출해 반환
 
     Args:
         text: 공고 제목 (예: "백엔드 개발자 채용")
 
     Returns:
-        np.ndarray: 768차원 임베딩 벡터
+        np.ndarray: 1024차원 임베딩 벡터
     """
 
-    tokenizer, model = KoBERTModel.get()
+    tokenizer, model = EmbeddingModel.get()
 
     inputs = tokenizer(
         text,
@@ -26,6 +26,10 @@ def extract_embedding(text: str) -> np.ndarray:
     with torch.no_grad():
         outputs = model(**inputs)
 
-    embedding = outputs.last_hidden_state[:, 0, :].squeeze().numpy()
+    # KURE-v1은 mean pooling 사용
+    attention_mask = inputs["attention_mask"]
+    token_embeddings = outputs.last_hidden_state
+    expanded_mask = attention_mask.unsqueeze(-1).expand(token_embeddings.size()).float()
+    embedding = (torch.sum(token_embeddings * expanded_mask, dim=1) / expanded_mask.sum(dim=1).clamp(min=1e-9)).squeeze().numpy()
 
     return embedding
