+++ complaint-orchestrator/src/main/resources/db/migration/V14__create_feature_embedding_table.sql
@@ -3,5 +3,5 @@ CREATE EXTENSION IF NOT EXISTS vector;
 CREATE TABLE feature_embedding (
    id           BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
    feature_id   BIGINT NOT NULL UNIQUE REFERENCES feature(id) ON DELETE CASCADE,
-   embedding    vector(768)
+   embedding    vector(1024)
 );
