+++ complaint-orchestrator/src/main/resources/db/migration/V15__create_process_code_type_embedding.sql
@@ -1,5 +1,5 @@
 CREATE TABLE process_code_type_embedding (
     id                      BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
     process_code_type_id    INT NOT NULL UNIQUE REFERENCES process_code_type(code) ON DELETE CASCADE,
-    embedding               vector(768)
+    embedding               vector(1024)
 );
