+++ complaint-orchestrator/src/main/resources/db/migration/V16__create_complaint_embedding_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE complaint_embedding (
      id             BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
      complaint_id   BIGINT NOT NULL UNIQUE REFERENCES complaint(id) ON DELETE CASCADE,
-     embedding      vector(768)
+     embedding      vector(1024)
 );
